requirements.txt
Django>=5.2.6
django-filter>=23.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
reportlab>=4.0.0
